Create publications page
Publications should be contained in the publications.yml file in the
data folder. The publications.html page will process the data and list
them either as published, submitted or in preparation. An
authorlists.html file in the include folder will process author lists,
to appropriately inject commas and the word 'and' as needed. It will
also truncate lists of more than five authors after the fourth offer and
append 'et al.'